Disable by default log files saving
If --output_dir command line options is omitted then
gtest-parallel does not save any log files at all.
Temporary files used for logging in this case are
removed immediately after test's exit. Logs for
failing tests are still dumped to console.

+++ gtest_parallel.py
@@ -27,7 +27,6 @@ import tempfile
 import thread
 import threading
 import time
-import zlib
 
 if sys.platform == 'win32':
   import msvcrt
@@ -185,6 +184,11 @@ class Task(object):
 
   @staticmethod
   def _logname(output_dir, test_binary, test_name, execution_number):
+    if output_dir is None:
+      (log_handle, log_name) = tempfile.mkstemp(prefix='tmpGP_', suffix=".log")
+      os.close(log_handle)
+      return log_name
+
     log_name = '%s-%s-%d.log' % (Task._normalize(os.path.basename(test_binary)),
                                  Task._normalize(test_name), execution_number)
 
@@ -290,6 +294,9 @@ class FilterFormat(object):
     self.stdout_lock = threading.Lock()
 
   def move_to(self, destination_dir, tasks):
+    if self.output_dir is None:
+      return
+
     destination_dir = os.path.join(self.output_dir, destination_dir)
     os.makedirs(destination_dir)
     for task in tasks:
@@ -320,6 +327,9 @@ class FilterFormat(object):
           "[%d/%d] %s returned/aborted with exit code %d (%d ms)"
           % (self.finished_tasks, self.total_tasks, task.test_name,
              task.exit_code, task.runtime_ms))
+
+    if self.output_dir is None:
+      os.remove(task.log_file)
 
   def log_tasks(self, total_tasks):
     self.total_tasks += total_tasks
@@ -624,8 +634,7 @@ def default_options_parser():
   parser = optparse.OptionParser(
       usage = 'usage: %prog [options] binary [binary ...] -- [additional args]')
 
-  parser.add_option('-d', '--output_dir', type='string',
-                    default=tempfile.gettempdir(),
+  parser.add_option('-d', '--output_dir', type='string', default=None,
                     help='Output directory for test logs. Logs will be '
                          'available under gtest-parallel-logs/, so '
                          '--output_dir=/tmp will results in all logs being '
@@ -682,7 +691,9 @@ def main():
   # data if an user passes a directory where files are already present. If a
   # user specifies --output_dir=Docs/, we'll create Docs/gtest-parallel-logs
   # and clean that directory out on startup, instead of nuking Docs/.
-  options.output_dir = os.path.join(options.output_dir, 'gtest-parallel-logs')
+  if options.output_dir is not None:
+    options.output_dir = os.path.join(options.output_dir,
+                                      'gtest-parallel-logs')
 
   if binaries == []:
     parser.print_usage()
@@ -703,16 +714,17 @@ def main():
   assert len(unique_binaries) == len(binaries), (
       "All test binaries must have an unique basename.")
 
-  # Remove files from old test runs.
-  if os.path.isdir(options.output_dir):
-    shutil.rmtree(options.output_dir)
-  # Create directory for test log output.
-  try:
-    os.makedirs(options.output_dir)
-  except OSError as e:
-    # Ignore errors if this directory already exists.
-    if e.errno != errno.EEXIST or not os.path.isdir(options.output_dir):
-      raise e
+  if options.output_dir is not None:
+    # Remove files from old test runs.
+    if os.path.isdir(options.output_dir):
+      shutil.rmtree(options.output_dir)
+    # Create directory for test log output.
+    try:
+      os.makedirs(options.output_dir)
+    except OSError as e:
+      # Ignore errors if this directory already exists.
+      if e.errno != errno.EEXIST or not os.path.isdir(options.output_dir):
+        raise e
 
   timeout = None
   if options.timeout is not None:
